Changed to Graph based naming scheme (ie. nodes and edges)
