--- computeScores.sh
@@ -1,1 +1,0 @@
-python3 src/scripts/computeScores.py
