+++ src/classes/Calendar.py
@@ -3,7 +3,17 @@ from monthdelta import monthdelta
 
 class Calendar:
 
+    def paths(self, startDate, endDate):
+
+        paths = []
+        for date in Calendar.dates(startDate, endDate):
+            paths.append('data/%s/%s/*.csv' % (self.name, date))
+
+        return paths
+
+
     def dates(startDate, endDate):
+
         date = startDate
 
         endDatetime = datetime.strptime(endDate, "%Y-%m")
