--- src/classes/abstract/RawInput.py
@@ -1,5 +1,0 @@
-class RawInput:
-    
-    def ingest(self, date):
-        self.df = self.comments.dataFrame(self.context, date)
-        return self
