--- src/scripts/computeLinks.py
@@ -1,21 +1,0 @@
-import sys
-sys.path.insert(0, 'src/classes/abstract')
-sys.path.insert(0, 'src/classes/concrete')
-sys.path.insert(0, 'src/classes')
-from Comments import Comments
-from Calendar import Calendar
-from MyContext import MyContext
-from Links import Links
-
-comments = Comments( comment_url = "https://files.pushshift.io/reddit/comments", \
-                           comment_path = 'files.pushshift.io/reddit/comments', \
-                           s3BucketName = "romeosredditcomments" )
-
-context = MyContext().context()
-
-startDate = '2005-12'
-endDate = '2006-02'
-links = Links(context, comments)
-
-links.process(startDate, endDate)
-
