+++ src/scripts/computeNodes.py
